Auto merged
2004/07/12 14:37:13-07:00 mvista.com!sdake
When a processor left the membership, the next configuration would sometimes
cause a form token timeout.  While not particularly harmful, it was wasteful
and not part of the original design of the group messaging protocol.
There was some extra junk code that was added to workaround some other bug
that has since been fixed.
This junk code removed and now the form token never times out (woohoo).
Also removed some extra code that calculates the next ORF processor twice.
We only really need to do it once.

(Logical change 1.33)


git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@100 fd59a12c-fef9-0310-b244-a6a79926bd2f