make all priv_data_pt pointers non-"const"
I'd suspected this would be necessary, but until now hadn't
found a case in which it actually was.  That case:
totemconfig.c's totem_key_change_notify.
It has a priv_data_pt parameter that is used like this:

  struct totem_config *totem_config = priv_data_pt;

and totem_config is in turn passed as arg #2 to
totem_volatile_config_read, where it is decidedly non-const.

git grep -l 'const void \*priv' \
  |xargs perl -pi -e 's,const (void \*priv_data_pt),$1,'

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2041 fd59a12c-fef9-0310-b244-a6a79926bd2f