Merge trunk revision 2365: r2365 | honzaf | 2009-07-20 07:57:06 -0700 (Mon, 20 Jul 2009) | 15 lines
Cpg synchronization patch for conf change messages

The root of the theoretical problem is that cpg_join or cpg_leave
messages are being sent via the C apis between synchronization.  With
the current cpg, synchronization happens in confchg_fn, and then later
in cpg_sync_process.  cpg_sync_process is called much later after
confchg_fn and introduces a small probability of a window of time for
queued in totem (but not yet ordered by totem) for those cpg_join and
cpg_leave operations to interact with the synchronization process which
should happen in one atomic operation but currently is two distinct
operations.

This patch deletes confchg_fn and make sends joinlist/downlist
in cpg_sync_process.


git-svn-id: http://svn.fedorahosted.org/svn/corosync/branches/flatiron@2423 fd59a12c-fef9-0310-b244-a6a79926bd2f