Merge trunk revision 2404: r2404 | asalkeld | 2009-09-07 01:33:23 -0700 (Mon, 07 Sep 2009) | 16 lines
Remove a double unlock in logsys.

In logsys there are a number of XYZ_unlocked() functions that are called
like this:

       pthread_mutex_lock (&logsys_config_mutex);

       XYZ_unlocked();

       pthread_mutex_unlock (&logsys_config_mutex);

The XYZ_unlocked() functions should not do any (un)locking.
But _logsys_config_subsys_get_unlocked() unlocks if it finds the subsys.
This means that there is 1 lock and 2 unlocks.


git-svn-id: http://svn.fedorahosted.org/svn/corosync/branches/flatiron@2453 fd59a12c-fef9-0310-b244-a6a79926bd2f