Merge trunk revision 2409: r2409 | asalkeld | 2009-09-09 11:55:17 -0700 (Wed, 09 Sep 2009) | 11 lines
don't return the ip address when the mask is not found

Currently totemip_copy (boundto,&ipaddr) is called even when the
required mask is not found.

This patch changes the behavior to only copy the ipaddr when the
mask is found.

The current behavior makes debugging an incorrect config really
confusing.



git-svn-id: http://svn.fedorahosted.org/svn/corosync/branches/flatiron@2456 fd59a12c-fef9-0310-b244-a6a79926bd2f