Merge trunk revision 2510: r2510 | chrissie | 2009-10-06 05:57:35 -0700 (Tue, 06 Oct 2009) | 13 lines
This patche fixes a couple of small problems with votequorum:

- if a single node is booted with votequorum loaded then
   corosync-quorumtool shows zero nodes and no votes.
- votequorum doesn't always tell the main quorum module when a new node
has joined the cluster (principally itself. this bug is actually tied
into the above)

I've also added quorum to the default list of services. As quorum has
been decoupled from sync it will not interfere with normal operations as
it used to do and it makes more sense to have it there than not.



git-svn-id: http://svn.fedorahosted.org/svn/corosync/branches/flatiron@2525 fd59a12c-fef9-0310-b244-a6a79926bd2f