coroipcc: use malloc for path in service_connect
Coroipcc appropriately uses PATH_MAX sized variables for various data
structures handling files in the initialization of the client.  Due to
the use of 12 of these structures declared as stack variables, the
application stack balloons to over 12*4k. This is especially problematic
if threads are used by long running daemons to restart the connection
to corosync so as to be resilient in the face of system services
restarting (service corosync restart).

A simple alternative is to allocate temporary memory to avoid
requirements of large thread stacks.

Original patch by Dan Clark <2clarkd@gmail.com>

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>