Don't unlock mutex in different threads
corosync_timer_init used to lock timer lock and unlock was made in
prioritized_timer_thread. Even this behavior works on Linux, NetBSD is
more strict. Mutex can be unlocked only by thread which locked it.

Solution is to use condition and synchronize start of timer thread and
main thread.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>