quorumtool: fix display when nodenames or reverse lookup is missing
when requesting nodename in ring0_addr, and reverse dns lookup
is not available, quorumtool output becomes inconsistent.

quorum {
    provider: corosync_votequorum
}

nodelist {
        node {
                ring0_addr: fedora-master-node1
                nodeid: 1
        }
        node {
                ring0_addr: fedora-master-node2
                nodeid: 2
        }
}

[fabbione@daikengo corosync]$ nslookup fedora-master-node1
..
Address: 192.168.2.193

[fabbione@daikengo corosync]$ nslookup 192.168.2.193
..
** server can't find 193.2.168.192.in-addr.arpa.: NXDOMAIN

[root@fedora-master-node1 tools]# corosync-quorumtool -s
...

Membership information
----------------------
    Nodeid      Votes    Qdevice Name
         1          1         NR fedora-master-node1.int.fabbione.net
         2          1         NR 192.168.2.194

(similar on the other node)

With this patch, when nodelist is available, we simply return ring0_addr
name for a node, fixing the output to be:

Membership information
----------------------
    Nodeid      Votes    Qdevice Name
         1          1         NR fedora-master-node1
         2          1         NR fedora-master-node2

that also removes possible inconsistencies between nodename, FQDN
and so on, by using what user requested in corosync.conf

In case ring0_addr is not available or numeric output is requested,
behavior remains as before.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>
Reviewed-by: Jan Friesse <jfriesse@redhat.com>