Add support for debug level trace in config file
Because logsys uses 3-bits for log level encoded in rec, it's impossible
to add trace log level in clean way. Instead of that, we are using
recid of TRACE1 for trace messages. So if trace is allowed in
configuration file, we change old condition to log only LOGSYS_RECID_LOG
to log also LOGSYS_RECID_TRACE1.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Fabio M. Di Nitto <fdinitto@redhat.com>